Ch2_Communication_MQTT_API/04_API_REST_Django/sensors_api/views.py
from rest_framework import generics
from .models import SensorData
from .serializers import SensorDataSerializer

class SensorDataListCreateView(generics.ListCreateAPIView):
    queryset = SensorData.objects.all().order_by('-timestamp')
    serializer_class = SensorDataSerializer
