Ch2_Communication_MQTT_API/04_API_REST_Django/requirements.txt
django
djangorestframework
paho-mqtt
tabulate
